Add Input.Checks package
The package provides the Run procedure which executes all defined
(pre)checks on the scheduling configuration. If an error is detected, a
Validation_Error exception is raised.

The currently implemented check Fixed_CPU_In_Range validates that all
user-specified fixed CPU allocations are in the allowed range of
available CPUs.